move join into periodics
